pyproject.toml
[project]
name = "redis-pubsub-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "uvicorn>=0.34.1",
]
